Revert "fix: firebase adapter singleton"
