+++ front/app/api/auth/[...nextauth]/route.js
@@ -2,9 +2,37 @@ import NextAuth from 'next-auth';
 import GoogleProvider from 'next-auth/providers/google';
 
 import { FirestoreAdapter } from '@auth/firebase-adapter';
-import { getAdminDb } from '@/lib/firebase-admin-singleton'; // Adjust path
+import * as admin from 'firebase-admin';
+import { getFirestore } from 'firebase-admin/firestore'; // Keep this
 
-const db = getAdminDb(); // This will be called on import
+if (!admin.apps.length) {
+  const serviceAccountJsonString = process.env.FIREBASE_SERVICE_ACCOUNT_JSON;
+  if (!serviceAccountJsonString) {
+    throw new Error(
+      'FIREBASE_SERVICE_ACCOUNT_JSON environment variable is not set or empty.',
+    );
+  }
+  try {
+    const serviceAccount = JSON.parse(serviceAccountJsonString);
+    admin.initializeApp({
+      credential: admin.credential.cert(serviceAccount),
+      // Optionally, specify databaseURL if not automatically inferred or if you have multiple DBs
+      // databaseURL: `https://your-project-id.firebaseio.com`
+    });
+    console.log('Firebase Admin SDK initialized successfully.');
+  } catch (error) {
+    console.error('Error initializing Firebase Admin SDK:', error);
+    console.error(
+      'Check if FIREBASE_SERVICE_ACCOUNT_JSON is a valid JSON string.',
+    );
+    throw new Error(
+      'Failed to initialize Firebase Admin SDK. Ensure FIREBASE_SERVICE_ACCOUNT_JSON is set correctly.',
+    );
+  }
+}
+
+// You can still get the Firestore instance like this for the adapter
+const db = getFirestore();
 
 export const authOptions = {
   providers: [
